eatgo-admin-api/src/main/java/kr/com/djam/eatgo/application/RegionService.java
package kr.com.djam.eatgo.application;

import kr.com.djam.eatgo.domain.Region;
import kr.com.djam.eatgo.domain.RegionRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;

@Service
@Transactional
public class RegionService {

    @Autowired
    private RegionRepository regionRepository;

    public RegionService(RegionRepository regionRepository) {
        this.regionRepository = regionRepository;
    }

    public List<Region> getRegions() {
        return regionRepository.findAll();
    }

    public Region addRegion(String name) {
        Region region = Region.builder().name(name).build();
        System.out.println(region.getName());
        return regionRepository.save(region);
    }
}
